+++ src/App.vue
@@ -4,7 +4,6 @@ import ChatbotWidget from './components/ChatbotWidget.vue'
 
 <template>
   <div class="min-h-screen bg-gray-100">
-    <ChatbotWidget user-id="a3aa1086-ff4e-47c6-81c4-01a1da01723b" thread-id="thread_ePgFSxkFT0ft9I8XyTCMUhgl"
-      company-id="b475ce33-495c-492b-b84b-8f1fd3a20531" />
+    <ChatbotWidget app-id="Ottoo_Hr_1_Tutku" app-name="TutkuAI" />
   </div>
 </template>