+++ src/widget.ts
@@ -10,25 +10,23 @@ const initWidget = () => {
     const script = scripts[i]
     if (
       script &&
-      (script.hasAttribute('data-user-id') || script.src.includes('tutku-chatbot-widget'))
+      (script.hasAttribute('data-app-id') || script.src.includes('tutku-chatbot-widget'))
     ) {
       scriptTag = script as HTMLScriptElement
       break
     }
   }
 
-  const userId = scriptTag?.getAttribute('data-user-id') || ''
-  const threadId = scriptTag?.getAttribute('data-thread-id') || ''
-  const companyId = scriptTag?.getAttribute('data-company-id') || ''
+  const appId = scriptTag?.getAttribute('data-app-id') || ''
+  const appName = scriptTag?.getAttribute('data-app-name') || ''
 
   const widgetContainer = document.createElement('div')
   widgetContainer.id = 'tutku-ai-chatbot-widget'
   document.body.appendChild(widgetContainer)
 
   const app = createApp(ChatbotWidget, {
-    userId,
-    threadId,
-    companyId,
+    appId,
+    appName,
   })
   app.mount(widgetContainer)
 }